Rewrite LocalizedHelpers This rewrite now acts only as a generated proxy for the standard `Helpers` module.
The standard helpers, for localized routes, are now
named (using the `:as` option) to include the locale
name. So `user_path` becomes `user_en_path` and
`user_de_path` for example.

The helpers in `LocalizedHelper` are named in the
standard manner, like `user_path` and when called will
delegate to the appropriate standard helper depending
on the locale returned by `Cldr.get_locale/1`.

In the current implementation, any user provided `:as`
option for the route is overridden. The next commit
will honour any supplied `:as` option but still append
the locale in order to ensure the delegation works
as expected.