LocalizedRoutes retains the localized routes LocalizedRoutes defines a funtion __route__/0 the includes %Phoenix.Route{} structs for each of the define routes after localization.
This commit undoes the attempt to "unlocalize"
the routes in order a developer has full visibility
into the available routes with
`mix phx.routes MyApp.CLdr.LocalizedRoutes`.

Another commit is required to format the
routes in manner which makes it easier to
identify the routes associated with a given
locale.