Add localized verified routes Introduces a sigil_q that supports localized verified routes in Phoenix 1.7 and later.
The choice of ~q as the sigil is driven by two
factors:

1. Overriding ~p might lead to developer confusion
2. On US keywords, the `p` and `q` are at opposite
   ends of the top keyboard row giving some symmetry
   and a memory prompt.

Using localised verified routes is as simple as adding:

 use MyAppWeb.Router.VerifiedRoutes,
   router: MyApp.Router,
   endpoint: MyApp.Endpoint

This could be further consolidated by adding that line to
the `MyAppWeb` module under the `:verified_routes` key.

The code generated by sigil_q invokes sigil_p under the
covers. The generated code looks like this:

  # ~q"/users" generates:
  case MyApp.Cldr.get_locale().cldr_locale_name do
    :de -> ~p"/users_de"
    :en -> ~p"/users"
    :fr -> ~p"/users_fr"
  end