Support fragment/query in localized verified route This commit adds support for URLs that include fragments (the "#location" construct at the end of a path) and query strings (the part after any optional "?" at the end of a path).
Neither the fragment or query string is translated.
This is by design since these parameters would normally
be considered either part of the page markup (the
fragment) or the api definition (query params).

The is room to change the decision not to translate
these elements. One approach would be to add two
flags to the ~q sigil. "f" for "translate fragment"
and "q" for "translate query keys".